+++ lib/src/io/mobile_unity_widget_controller.dart
@@ -1,9 +1,16 @@
-part of flutter_unity_widget;
+import 'dart:async';
 
-typedef void UnityCreatedCallback(UnityWidgetController controller);
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
-class UnityWidgetController {
-  final _UnityWidgetState _unityWidgetState;
+import '../facade_controller.dart';
+import '../helpers/events.dart';
+import 'device_method.dart';
+import 'unity_widget.dart';
+import 'unity_widget_platform.dart';
+
+class MobileUnityWidgetController extends UnityWidgetController {
+  final MobileUnityWidgetState _unityWidgetState;
 
   /// The unityId for this controller
   final int unityId;
@@ -13,17 +20,18 @@ class UnityWidgetController {
       _onUnitySceneLoadedSub,
       _onUnityUnloadedSub;
 
-  UnityWidgetController._(this._unityWidgetState, {required this.unityId}) {
+  MobileUnityWidgetController._(this._unityWidgetState,
+      {required this.unityId}) {
     _connectStreams(unityId);
   }
 
   /// Initialize [UnityWidgetController] with [id]
   /// Mainly for internal use when instantiating a [UnityWidgetController] passed
   /// in [UnityWidget.onUnityCreated] callback.
-  static Future<UnityWidgetController> init(
-      int id, _UnityWidgetState unityWidgetState) async {
-    await UnityViewFlutterPlatform.instance.init(id);
-    return UnityWidgetController._(
+  static Future<MobileUnityWidgetController> init(
+      int id, MobileUnityWidgetState unityWidgetState) async {
+    await UnityWidgetPlatform.instance.init(id);
+    return MobileUnityWidgetController._(
       unityWidgetState,
       unityId: id,
     );
@@ -31,9 +39,8 @@ class UnityWidgetController {
 
   @visibleForTesting
   MethodChannel? get channel {
-    if (UnityViewFlutterPlatform.instance is MethodChannelUnityViewFlutter) {
-      return (UnityViewFlutterPlatform.instance
-              as MethodChannelUnityViewFlutter)
+    if (UnityWidgetPlatform.instance is MethodChannelUnityWidget) {
+      return (UnityWidgetPlatform.instance as MethodChannelUnityWidget)
           .channel(unityId);
     }
     return null;
@@ -41,21 +48,21 @@ class UnityWidgetController {
 
   void _connectStreams(int unityId) {
     if (_unityWidgetState.widget.onUnityMessage != null) {
-      _onUnityMessageSub = UnityViewFlutterPlatform.instance
+      _onUnityMessageSub = UnityWidgetPlatform.instance
           .onUnityMessage(unityId: unityId)
           .listen((UnityMessageEvent e) =>
               _unityWidgetState.widget.onUnityMessage!(e.value));
     }
 
     if (_unityWidgetState.widget.onUnitySceneLoaded != null) {
-      _onUnitySceneLoadedSub = UnityViewFlutterPlatform.instance
+      _onUnitySceneLoadedSub = UnityWidgetPlatform.instance
           .onUnitySceneLoaded(unityId: unityId)
           .listen((UnitySceneLoadedEvent e) =>
               _unityWidgetState.widget.onUnitySceneLoaded!(e.value));
     }
 
     if (_unityWidgetState.widget.onUnityUnloaded != null) {
-      _onUnityUnloadedSub = UnityViewFlutterPlatform.instance
+      _onUnityUnloadedSub = UnityWidgetPlatform.instance
           .onUnityUnloaded(unityId: unityId)
           .listen((_) => _unityWidgetState.widget.onUnityUnloaded!());
     }
@@ -65,7 +72,7 @@ class UnityWidgetController {
   /// Returns `true` if unity player is ready.
   Future<bool?>? isReady() {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance.isReady(unityId: unityId);
+      return UnityWidgetPlatform.instance.isReady(unityId: unityId);
     }
     return null;
   }
@@ -74,7 +81,7 @@ class UnityWidgetController {
   /// Returns `true` if unity player is paused.
   Future<bool?>? isPaused() {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance.isPaused(unityId: unityId);
+      return UnityWidgetPlatform.instance.isPaused(unityId: unityId);
     }
     return null;
   }
@@ -83,7 +90,7 @@ class UnityWidgetController {
   /// Returns `true` if unity player is loaded.
   Future<bool?>? isLoaded() {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance.isLoaded(unityId: unityId);
+      return UnityWidgetPlatform.instance.isLoaded(unityId: unityId);
     }
     return null;
   }
@@ -92,7 +99,7 @@ class UnityWidgetController {
   /// Returns `true` if unity player is in background.
   Future<bool?>? inBackground() {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance.inBackground(unityId: unityId);
+      return UnityWidgetPlatform.instance.inBackground(unityId: unityId);
     }
     return null;
   }
@@ -102,8 +109,7 @@ class UnityWidgetController {
   /// Returns `true` if unity player was created succesfully.
   Future<bool?>? create() {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance
-          .createUnityPlayer(unityId: unityId);
+      return UnityWidgetPlatform.instance.createUnityPlayer(unityId: unityId);
     }
     return null;
   }
@@ -117,7 +123,7 @@ class UnityWidgetController {
   /// ```
   Future<void>? postMessage(String gameObject, methodName, message) {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance.postMessage(
+      return UnityWidgetPlatform.instance.postMessage(
         unityId: unityId,
         gameObject: gameObject,
         methodName: methodName,
@@ -137,7 +143,7 @@ class UnityWidgetController {
   Future<void>? postJsonMessage(
       String gameObject, String methodName, Map<String, dynamic> message) {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance.postJsonMessage(
+      return UnityWidgetPlatform.instance.postJsonMessage(
         unityId: unityId,
         gameObject: gameObject,
         methodName: methodName,
@@ -150,7 +156,7 @@ class UnityWidgetController {
   /// Pause the unity in-game player with this method
   Future<void>? pause() {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance.pausePlayer(unityId: unityId);
+      return UnityWidgetPlatform.instance.pausePlayer(unityId: unityId);
     }
     return null;
   }
@@ -158,7 +164,7 @@ class UnityWidgetController {
   /// Resume the unity in-game player with this method idf it is in a paused state
   Future<void>? resume() {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance.resumePlayer(unityId: unityId);
+      return UnityWidgetPlatform.instance.resumePlayer(unityId: unityId);
     }
     return null;
   }
@@ -167,8 +173,7 @@ class UnityWidgetController {
   /// It works for Android and iOS is WIP
   Future<void>? openInNativeProcess() {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance
-          .openInNativeProcess(unityId: unityId);
+      return UnityWidgetPlatform.instance.openInNativeProcess(unityId: unityId);
     }
     return null;
   }
@@ -177,7 +182,7 @@ class UnityWidgetController {
   /// iOS is WIP. For more information please read [Unity Docs](https://docs.unity3d.com/2020.2/Documentation/Manual/UnityasaLibrary.html)
   Future<void>? unload() {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance.unloadPlayer(unityId: unityId);
+      return UnityWidgetPlatform.instance.unloadPlayer(unityId: unityId);
     }
     return null;
   }
@@ -185,7 +190,7 @@ class UnityWidgetController {
   /// Quits unity player. Note that this kills the current flutter process, thus quiting the app
   Future<void>? quit() {
     if (!_unityWidgetState.widget.enablePlaceholder) {
-      return UnityViewFlutterPlatform.instance.quitPlayer(unityId: unityId);
+      return UnityWidgetPlatform.instance.quitPlayer(unityId: unityId);
     }
     return null;
   }
@@ -203,12 +208,6 @@ class UnityWidgetController {
 
   void dispose() {
     _cancelSubscriptions();
-    UnityViewFlutterPlatform.instance.dispose(unityId: unityId);
+    UnityWidgetPlatform.instance.dispose(unityId: unityId);
   }
 }
-
-typedef void UnityMessageCallback(dynamic handler);
-
-typedef void UnitySceneChangeCallback(SceneLoaded? message);
-
-typedef void UnityUnloadCallback();